Added in some of the art, changed map generation
Added in some of the art into the game (mostly the UI elements), and
changed the map generation so that there is a larger island in the
middle of the map in case we want to add any game elements such as land
combat or events. Will work more tomorrow.